refactor(camera): factor shared driver logic into _AbstractUvcCamera
Move ArucoState (formerly _ArucoState) and _rot_to_quat to perception/aruco.py
since they have no V4L2 dependency. Introduce _AbstractUvcCamera as the
shared base for UvcCamera and UvcCameraVirtual: subclasses now only override
init/close/capture/set_focus/set_setting plus a _init_state hook for their
private state. The signature-parity invariant is now enforced by inheritance
instead of by discipline.

Also mutualize UvcCameraDefinition / UvcCameraVirtualDefinition via a
shared base class and a _add_camera_args helper, so adding a new
constructor param now only touches one place.
