+++ src/evo_lib/drivers/camera/uvc.py
@@ -25,13 +25,7 @@ from evo_lib.interfaces.camera import (
     Camera,
 )
 from evo_lib.logger import Logger
-from evo_lib.perception.aruco import (
-    CharucoCalibrationSession,
-    _aruco_dict_from_name,
-    _build_charuco_board,
-    _marker_object_points,
-    _now_iso,
-)
+from evo_lib.perception.aruco import ArucoState, CharucoCalibrationSession
 from evo_lib.task import ImmediateResultTask, Task
 
 if TYPE_CHECKING:
@@ -62,258 +56,9 @@ def _v4l2_has_ctrl(device: str, ctrl: str) -> bool:
     return result.returncode == 0 and ctrl in result.stdout
 
 
-def _rot_to_quat(R: "np.ndarray") -> tuple[float, float, float, float]:
-    """3x3 rotation matrix to (qw, qx, qy, qz) unit quaternion (Shepperd's method)."""
-    import numpy as np
+class _AbstractUvcCamera(Camera):
+    """Shared layer for UVC camera drivers; subclasses provide the V4L2 ops."""
 
-    trace = R[0, 0] + R[1, 1] + R[2, 2]
-    if trace > 0.0:
-        s = 2.0 * np.sqrt(1.0 + trace)
-        return (
-            float(0.25 * s),
-            float((R[2, 1] - R[1, 2]) / s),
-            float((R[0, 2] - R[2, 0]) / s),
-            float((R[1, 0] - R[0, 1]) / s),
-        )
-    if R[0, 0] > R[1, 1] and R[0, 0] > R[2, 2]:
-        s = 2.0 * np.sqrt(1.0 + R[0, 0] - R[1, 1] - R[2, 2])
-        return (
-            float((R[2, 1] - R[1, 2]) / s),
-            float(0.25 * s),
-            float((R[0, 1] + R[1, 0]) / s),
-            float((R[0, 2] + R[2, 0]) / s),
-        )
-    if R[1, 1] > R[2, 2]:
-        s = 2.0 * np.sqrt(1.0 + R[1, 1] - R[0, 0] - R[2, 2])
-        return (
-            float((R[0, 2] - R[2, 0]) / s),
-            float((R[0, 1] + R[1, 0]) / s),
-            float(0.25 * s),
-            float((R[1, 2] + R[2, 1]) / s),
-        )
-    s = 2.0 * np.sqrt(1.0 + R[2, 2] - R[0, 0] - R[1, 1])
-    return (
-        float((R[1, 0] - R[0, 1]) / s),
-        float((R[0, 2] + R[2, 0]) / s),
-        float((R[1, 2] + R[2, 1]) / s),
-        float(0.25 * s),
-    )
-
-
-class _ArucoState:
-    """Detection + calibration state shared between UvcCamera and UvcCameraVirtual."""
-
-    def __init__(
-        self,
-        marker_size_mm: float,
-        dictionary: str,
-        charuco_squares_x: int,
-        charuco_squares_y: int,
-        charuco_square_mm: float,
-        charuco_marker_mm: float,
-        calibration_path: str | None,
-    ):
-        self.marker_size_mm = marker_size_mm
-        self.dictionary_name = dictionary
-        self.charuco_geom = (
-            charuco_squares_x,
-            charuco_squares_y,
-            charuco_square_mm,
-            charuco_marker_mm,
-        )
-        self.calibration_path = calibration_path
-        self.K = None
-        self.dist = None
-        self.image_size: tuple[int, int] | None = None
-        self.R_robot_camera = None
-        self.t_robot_camera = None
-        self._obj_pts_cache: dict[float, "np.ndarray"] = {
-            marker_size_mm: _marker_object_points(marker_size_mm)
-        }
-        self._detector_handle = None
-
-    def _obj_pts_for(self, size_mm: float) -> "np.ndarray":
-        cached = self._obj_pts_cache.get(size_mm)
-        if cached is None:
-            cached = _marker_object_points(size_mm)
-            self._obj_pts_cache[size_mm] = cached
-        return cached
-
-    def _size_for_id(self, marker_id: int) -> float:
-        from evo_lib.types import EUROBOT_TAG_SIZES_MM
-
-        return EUROBOT_TAG_SIZES_MM.get(marker_id, self.marker_size_mm)
-
-    def detector(self) -> Any:
-        import cv2
-
-        if self._detector_handle is None:
-            dictionary = _aruco_dict_from_name(self.dictionary_name)
-            self._detector_handle = cv2.aruco.ArucoDetector(
-                dictionary, cv2.aruco.DetectorParameters()
-            )
-        return self._detector_handle
-
-    def detect_in_frame(self, frame: "np.ndarray") -> list[ArucoMarker]:
-        import cv2
-        import numpy as np
-
-        gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY) if frame.ndim == 3 else frame
-        marker_corners, marker_ids, _ = self.detector().detectMarkers(gray)
-        if marker_ids is None or len(marker_ids) == 0:
-            return []
-
-        results: list[ArucoMarker] = []
-        for i, marker_id in enumerate(marker_ids.flatten()):
-            corners_ij = marker_corners[i].reshape(4, 2).astype(np.float32)
-            mid = int(marker_id)
-            marker = ArucoMarker(id=mid, corners=corners_ij)
-            if self.K is not None and self.dist is not None:
-                obj_pts = self._obj_pts_for(self._size_for_id(mid))
-                ok, rvec, tvec = cv2.solvePnP(
-                    obj_pts,
-                    corners_ij,
-                    self.K,
-                    self.dist,
-                    flags=cv2.SOLVEPNP_IPPE_SQUARE,
-                )
-                if ok:
-                    marker.rvec_camera = rvec.reshape(3)
-                    marker.tvec_camera = tvec.reshape(3)
-                    if self.R_robot_camera is not None and self.t_robot_camera is not None:
-                        pos_robot = self.R_robot_camera @ marker.tvec_camera.reshape(
-                            3, 1
-                        ) + self.t_robot_camera.reshape(3, 1)
-                        marker.position_robot_mm = pos_robot.reshape(3)
-                        R_camera_marker, _ = cv2.Rodrigues(rvec)
-                        R_robot_marker = self.R_robot_camera @ R_camera_marker
-                        qw, qx, qy, qz = _rot_to_quat(R_robot_marker)
-                        marker.quat_robot = (qw, qx, qy, qz)
-                        siny = 2.0 * (qw * qz + qx * qy)
-                        cosy = 1.0 - 2.0 * (qy * qy + qz * qz)
-                        marker.yaw_robot_rad = float(np.arctan2(siny, cosy))
-            results.append(marker)
-        return results
-
-    def calibration_session(self) -> CharucoCalibrationSession:
-        sx, sy, sq_mm, mk_mm = self.charuco_geom
-        board = _build_charuco_board(sx, sy, sq_mm, mk_mm, self.dictionary_name)
-        dictionary = _aruco_dict_from_name(self.dictionary_name)
-        return CharucoCalibrationSession(board=board, dictionary=dictionary)
-
-    def apply_intrinsics(self, intrinsics: dict[str, Any]) -> None:
-        import numpy as np
-
-        self.K = np.asarray(intrinsics["K"], dtype=np.float64).reshape(3, 3)
-        self.dist = np.asarray(intrinsics["dist"], dtype=np.float64).reshape(-1)
-        self.image_size = tuple(intrinsics["image_size"])
-
-    def is_intrinsics_loaded(self) -> bool:
-        return self.K is not None and self.dist is not None
-
-    def is_extrinsics_loaded(self) -> bool:
-        return self.R_robot_camera is not None and self.t_robot_camera is not None
-
-    def calibrate_extrinsics(
-        self,
-        markers: list[ArucoMarker],
-        reference_marker_id: int,
-        R_robot_marker: "np.ndarray",
-        t_robot_marker_mm: "np.ndarray",
-    ) -> dict[str, Any]:
-        # T_robot_camera = T_robot_marker @ inv(T_camera_marker).
-        # Rotation matrices are orthonormal so R^{-1} = R^T (no expensive inverse).
-        import cv2
-        import numpy as np
-
-        if self.K is None or self.dist is None:
-            raise RuntimeError("Cannot calibrate extrinsics without intrinsics first.")
-
-        target = next((m for m in markers if m.id == reference_marker_id), None)
-        if target is None or target.rvec_camera is None or target.tvec_camera is None:
-            raise RuntimeError(
-                f"Reference marker {reference_marker_id} not visible "
-                f"(detected ids: {[m.id for m in markers]})"
-            )
-
-        R_camera_marker, _ = cv2.Rodrigues(target.rvec_camera.reshape(3, 1))
-        t_camera_marker = target.tvec_camera.reshape(3, 1)
-        R_marker_camera = R_camera_marker.T
-        t_marker_camera = -R_marker_camera @ t_camera_marker
-
-        R_robot_camera = np.asarray(R_robot_marker) @ R_marker_camera
-        t_robot_camera = np.asarray(R_robot_marker) @ t_marker_camera + np.asarray(
-            t_robot_marker_mm
-        ).reshape(3, 1)
-
-        self.R_robot_camera = R_robot_camera
-        self.t_robot_camera = t_robot_camera
-
-        return {
-            "R_robot_camera": R_robot_camera.flatten().tolist(),
-            "t_robot_camera_mm": t_robot_camera.flatten().tolist(),
-            "reference_marker_id": int(reference_marker_id),
-            "captured_at": _now_iso(),
-        }
-
-    def load(self, path: str | None = None) -> bool:
-        target = path or self.calibration_path
-        if target is None or not os.path.exists(target):
-            return False
-
-        import json5
-        import numpy as np
-
-        with open(target, "r") as f:
-            data = json5.load(f)
-
-        intr = data.get("intrinsics")
-        if intr is not None:
-            self.K = np.asarray(intr["K"], dtype=np.float64).reshape(3, 3)
-            self.dist = np.asarray(intr["dist"], dtype=np.float64).reshape(-1)
-            self.image_size = tuple(intr.get("image_size", (0, 0)))
-
-        extr = data.get("extrinsics")
-        if extr is not None:
-            # Legacy json5 used R_world_camera / t_world_camera_mm — accept both
-            # so robots calibrated before the rename keep working without
-            # forcing a recalibration.
-            R_key = "R_robot_camera" if "R_robot_camera" in extr else "R_world_camera"
-            t_key = "t_robot_camera_mm" if "t_robot_camera_mm" in extr else "t_world_camera_mm"
-            self.R_robot_camera = np.asarray(extr[R_key], dtype=np.float64).reshape(3, 3)
-            self.t_robot_camera = np.asarray(extr[t_key], dtype=np.float64).reshape(3)
-        return intr is not None or extr is not None
-
-    def save(self, path: str | None = None) -> str:
-        target = path or self.calibration_path
-        if target is None:
-            raise RuntimeError("no calibration_path configured")
-
-        import json5
-
-        parent = os.path.dirname(target)
-        if parent:
-            os.makedirs(parent, exist_ok=True)
-        payload: dict[str, Any] = {}
-        if self.K is not None and self.dist is not None:
-            payload["intrinsics"] = {
-                "K": self.K.flatten().tolist(),
-                "dist": self.dist.flatten().tolist(),
-                "image_size": list(self.image_size or (0, 0)),
-                "captured_at": _now_iso(),
-            }
-        if self.R_robot_camera is not None and self.t_robot_camera is not None:
-            payload["extrinsics"] = {
-                "R_robot_camera": self.R_robot_camera.flatten().tolist(),
-                "t_robot_camera_mm": self.t_robot_camera.flatten().tolist(),
-                "captured_at": _now_iso(),
-            }
-        with open(target, "w") as f:
-            json5.dump(payload, f, indent=2)
-        return target
-
-
-class UvcCamera(Camera):
     commands = DriverCommands(parents=[Camera.commands])
 
     def __init__(
@@ -344,15 +89,7 @@ class UvcCamera(Camera):
         self._focus = focus
         self._autofocus = autofocus
         self._settings = dict(settings) if settings else {}
-        self._cap = None
-        self._lock = threading.Lock()
-        # Background reader: see frame_capture.md. Grab+retrieve in a loop
-        # (= cv2.VideoCapture.read), keep only the latest frame for callers.
-        self._read_thread: threading.Thread | None = None
-        self._read_stop = threading.Event()
-        self._latest_frame: "np.ndarray | None" = None
-        self._latest_frame_lock = threading.Lock()
-        self._aruco = _ArucoState(
+        self._aruco = ArucoState(
             marker_size_mm=marker_size_mm,
             dictionary=dictionary,
             charuco_squares_x=charuco_squares_x,
@@ -361,13 +98,86 @@ class UvcCamera(Camera):
             charuco_marker_mm=charuco_marker_mm,
             calibration_path=calibration_path,
         )
+        self._init_state()
+
+    def _init_state(self) -> None:
+        """Hook for subclasses to initialize their private V4L2/replay state."""
+
+    def init(self) -> Task[()]:
+        raise NotImplementedError
+
+    def close(self) -> None:
+        raise NotImplementedError
+
+    def capture(self) -> "np.ndarray":
+        raise NotImplementedError
+
+    def set_focus(self, value: int) -> Task[()]:
+        raise NotImplementedError
+
+    def set_setting(self, name: str, value: int) -> Task[()]:
+        raise NotImplementedError
+
+    def detect(self) -> list[ArucoMarker]:
+        return self._aruco.detect_in_frame(self.capture())
+
+    def calibration_session(self) -> CharucoCalibrationSession:
+        return self._aruco.calibration_session()
+
+    def apply_intrinsics(self, intrinsics: dict[str, Any]) -> None:
+        self._aruco.apply_intrinsics(intrinsics)
+
+    def calibrate_extrinsics(
+        self,
+        reference_marker_id: int,
+        R_robot_marker: "np.ndarray",
+        t_robot_marker_mm: "np.ndarray",
+    ) -> dict[str, Any]:
+        return self._aruco.calibrate_extrinsics(
+            self.detect(), reference_marker_id, R_robot_marker, t_robot_marker_mm
+        )
+
+    def is_intrinsics_loaded(self) -> bool:
+        return self._aruco.is_intrinsics_loaded()
+
+    def is_extrinsics_loaded(self) -> bool:
+        return self._aruco.is_extrinsics_loaded()
+
+    def save_calibration(self, path: str | None = None) -> None:
+        self._aruco.save(path)
+
+    @property
+    def K(self) -> "np.ndarray | None":
+        return self._aruco.K
+
+    @property
+    def dist(self) -> "np.ndarray | None":
+        return self._aruco.dist
+
+    @property
+    def width(self) -> int:
+        return self._width
+
+    @property
+    def height(self) -> int:
+        return self._height
+
+
+class UvcCamera(_AbstractUvcCamera):
+    def _init_state(self) -> None:
+        self._cap = None
+        self._lock = threading.Lock()
+        # Background reader: see frame_capture.md.
+        self._read_thread: threading.Thread | None = None
+        self._read_stop = threading.Event()
+        self._latest_frame: "np.ndarray | None" = None
+        self._latest_frame_lock = threading.Lock()
 
     def init(self) -> Task[()]:
         import cv2
 
         # AF must be locked before VideoCapture grabs its first buffer,
         # otherwise focal length floats and any K calibrated downstream is invalid.
-        # Fixed-focus lenses don't expose this control at all (e.g. U20CAM-1080p).
         if _v4l2_has_ctrl(self._device, "focus_automatic_continuous"):
             af_value = 1 if self._autofocus else 0
             ok, err = _v4l2_set_ctrl(self._device, "focus_automatic_continuous", af_value)
@@ -395,8 +205,7 @@ class UvcCamera(Camera):
         if not self._cap.isOpened():
             raise RuntimeError(f"UvcCamera '{self.name}': failed to open {self._device}")
 
-        # FOURCC must be set before the size: UVC's resolution table is
-        # format-dependent, MJPG and YUYV expose different size lists.
+        # FOURCC before size: UVC's resolution table is format-dependent.
         self._cap.set(cv2.CAP_PROP_FOURCC, cv2.VideoWriter_fourcc(*self._fourcc))
         self._cap.set(cv2.CAP_PROP_FRAME_WIDTH, self._width)
         self._cap.set(cv2.CAP_PROP_FRAME_HEIGHT, self._height)
@@ -422,7 +231,6 @@ class UvcCamera(Camera):
         return ImmediateResultTask()
 
     def _read_loop(self) -> None:
-        """Read frames continuously and keep only the latest. See frame_capture.md."""
         while not self._read_stop.is_set():
             with self._lock:
                 cap = self._cap
@@ -454,7 +262,6 @@ class UvcCamera(Camera):
     def capture(self) -> "np.ndarray":
         if self._cap is None:
             raise RuntimeError(f"UvcCamera '{self.name}' not opened")
-        # Wait briefly for the background thread to publish its first frame.
         deadline = time.monotonic() + 1.0
         while True:
             with self._latest_frame_lock:
@@ -464,9 +271,6 @@ class UvcCamera(Camera):
             if time.monotonic() > deadline:
                 raise RuntimeError(f"UvcCamera '{self.name}': no frame available")
             time.sleep(0.005)
-
-    def detect(self) -> list[ArucoMarker]:
-        return self._aruco.detect_in_frame(self.capture())
 
     def set_focus(self, value: int) -> Task[()]:
         ok, err = _v4l2_set_ctrl(self._device, "focus_absolute", value)
@@ -484,173 +288,22 @@ class UvcCamera(Camera):
             self._settings[name] = value
         return ImmediateResultTask()
 
-    def calibration_session(self) -> CharucoCalibrationSession:
-        return self._aruco.calibration_session()
-
-    def apply_intrinsics(self, intrinsics: dict[str, Any]) -> None:
-        self._aruco.apply_intrinsics(intrinsics)
-
-    def calibrate_extrinsics(
-        self,
-        reference_marker_id: int,
-        R_robot_marker: "np.ndarray",
-        t_robot_marker_mm: "np.ndarray",
-    ) -> dict[str, Any]:
-        return self._aruco.calibrate_extrinsics(
-            self.detect(), reference_marker_id, R_robot_marker, t_robot_marker_mm
-        )
-
-    def calibrate_extrinsics_from_pose(
-        self,
-        reference_marker_id: int,
-        rvec_camera_marker: "np.ndarray",
-        tvec_camera_marker: "np.ndarray",
-        R_robot_marker: "np.ndarray",
-        t_robot_marker_mm: "np.ndarray",
-    ) -> dict[str, Any]:
-        # Apply extrinsics from a pre-computed marker pose (e.g. averaged
-        # across N frames via bundle PnP) instead of running detect() again.
-        import numpy as np
-
-        marker = ArucoMarker(
-            id=reference_marker_id,
-            corners=np.zeros((4, 2), dtype=np.float32),
-        )
-        marker.rvec_camera = np.asarray(rvec_camera_marker).reshape(3)
-        marker.tvec_camera = np.asarray(tvec_camera_marker).reshape(3)
-        return self._aruco.calibrate_extrinsics(
-            [marker], reference_marker_id, R_robot_marker, t_robot_marker_mm
-        )
-
-    @property
-    def K(self) -> "np.ndarray | None":
-        return self._aruco.K
-
-    @property
-    def dist(self) -> "np.ndarray | None":
-        return self._aruco.dist
-
-    def detect_from_frame(self, frame: "np.ndarray") -> list[ArucoMarker]:
-        return self._aruco.detect_in_frame(frame)
-
-    def is_intrinsics_loaded(self) -> bool:
-        return self._aruco.is_intrinsics_loaded()
-
-    def is_extrinsics_loaded(self) -> bool:
-        return self._aruco.is_extrinsics_loaded()
-
-    def load_calibration(self, path: str | None = None) -> bool:
-        return self._aruco.load(path)
-
     def save_calibration(self, path: str | None = None) -> None:
         target = self._aruco.save(path)
         self._log.info(f"UvcCamera '{self.name}' saved calibration -> {target}")
 
-    @property
-    def width(self) -> int:
-        return self._width
 
-    @property
-    def height(self) -> int:
-        return self._height
-
-
-class UvcCameraDefinition(DriverDefinition):
-    def __init__(self, logger: Logger):
-        super().__init__(UvcCamera.commands)
-        self._logger = logger
-
-    def get_init_args_definition(self) -> DriverInitArgsDefinition:
-        defn = DriverInitArgsDefinition()
-        defn.add_required("device", ArgTypes.String())
-        defn.add_optional("width", ArgTypes.U32(), DEFAULT_WIDTH)
-        defn.add_optional("height", ArgTypes.U32(), DEFAULT_HEIGHT)
-        defn.add_optional("fourcc", ArgTypes.String(), DEFAULT_FOURCC)
-        defn.add_optional("focus", ArgTypes.I32(), -1)  # -1 = no override
-        defn.add_optional("autofocus", ArgTypes.Bool(), False)
-        defn.add_optional("marker_size_mm", ArgTypes.F32(), 32.0)
-        defn.add_optional("dictionary", ArgTypes.String(), DEFAULT_ARUCO_DICTIONARY)
-        defn.add_optional("calibration_path", ArgTypes.String(), "")
-        defn.add_optional("charuco_squares_x", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_X)
-        defn.add_optional("charuco_squares_y", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_Y)
-        defn.add_optional("charuco_square_mm", ArgTypes.F32(), DEFAULT_CHARUCO_SQUARE_MM)
-        defn.add_optional("charuco_marker_mm", ArgTypes.F32(), DEFAULT_CHARUCO_MARKER_MM)
-        return defn
-
-    def create(self, args: DriverInitArgs) -> UvcCamera:
-        focus_arg = args.get("focus")
-        focus = None if focus_arg == -1 else int(focus_arg)
-        cal_path = args.get("calibration_path") or None
-        return UvcCamera(
-            name=args.get_name(),
-            logger=self._logger,
-            device=args.get("device"),
-            width=args.get("width"),
-            height=args.get("height"),
-            fourcc=args.get("fourcc"),
-            focus=focus,
-            autofocus=args.get("autofocus"),
-            settings=None,
-            marker_size_mm=args.get("marker_size_mm"),
-            dictionary=args.get("dictionary"),
-            calibration_path=cal_path,
-            charuco_squares_x=args.get("charuco_squares_x"),
-            charuco_squares_y=args.get("charuco_squares_y"),
-            charuco_square_mm=args.get("charuco_square_mm"),
-            charuco_marker_mm=args.get("charuco_marker_mm"),
-        )
-
-
-class UvcCameraVirtual(Camera):
+class UvcCameraVirtual(_AbstractUvcCamera):
     """Drop-in replacement for UvcCamera. ``device`` read as a folder of
     JPEG/PNG to replay, or ignored to yield blank frames. Same ArUco surface."""
 
-    commands = DriverCommands(parents=[Camera.commands])
-
-    def __init__(
-        self,
-        name: str,
-        logger: Logger,
-        device: str,
-        width: int = DEFAULT_WIDTH,
-        height: int = DEFAULT_HEIGHT,
-        fourcc: str = DEFAULT_FOURCC,
-        focus: int | None = None,
-        autofocus: bool = False,
-        settings: dict[str, int] | None = None,
-        marker_size_mm: float = 32.0,
-        dictionary: str = DEFAULT_ARUCO_DICTIONARY,
-        calibration_path: str | None = None,
-        charuco_squares_x: int = DEFAULT_CHARUCO_SQUARES_X,
-        charuco_squares_y: int = DEFAULT_CHARUCO_SQUARES_Y,
-        charuco_square_mm: float = DEFAULT_CHARUCO_SQUARE_MM,
-        charuco_marker_mm: float = DEFAULT_CHARUCO_MARKER_MM,
-    ):
-        super().__init__(name)
-        self._log = logger
-        self._device = device
-        self._width = width
-        self._height = height
-        # fourcc / focus / autofocus kept for signature parity with UvcCamera
-        self._fourcc = fourcc
-        self._focus = focus
-        self._autofocus = autofocus
-        self._settings = dict(settings) if settings else {}
+    def _init_state(self) -> None:
         self._frames: list[str] = []
         self._injected_frames: list["np.ndarray"] | None = None
         self._injected_markers: list[ArucoMarker] | None = None
         self._cursor = 0
         self._lock = threading.Lock()
         self._opened = False
-        self._aruco = _ArucoState(
-            marker_size_mm=marker_size_mm,
-            dictionary=dictionary,
-            charuco_squares_x=charuco_squares_x,
-            charuco_squares_y=charuco_squares_y,
-            charuco_square_mm=charuco_square_mm,
-            charuco_marker_mm=charuco_marker_mm,
-            calibration_path=calibration_path,
-        )
 
     def init(self) -> Task[()]:
         if os.path.isdir(self._device):
@@ -700,7 +353,7 @@ class UvcCameraVirtual(Camera):
     def detect(self) -> list[ArucoMarker]:
         if self._injected_markers is not None:
             return list(self._injected_markers)
-        return self._aruco.detect_in_frame(self.capture())
+        return super().detect()
 
     def set_focus(self, value: int) -> Task[()]:
         self._focus = value
@@ -709,75 +362,6 @@ class UvcCameraVirtual(Camera):
     def set_setting(self, name: str, value: int) -> Task[()]:
         self._settings[name] = value
         return ImmediateResultTask()
-
-    def calibration_session(self) -> CharucoCalibrationSession:
-        return self._aruco.calibration_session()
-
-    def apply_intrinsics(self, intrinsics: dict[str, Any]) -> None:
-        self._aruco.apply_intrinsics(intrinsics)
-
-    def calibrate_extrinsics(
-        self,
-        reference_marker_id: int,
-        R_robot_marker: "np.ndarray",
-        t_robot_marker_mm: "np.ndarray",
-    ) -> dict[str, Any]:
-        return self._aruco.calibrate_extrinsics(
-            self.detect(), reference_marker_id, R_robot_marker, t_robot_marker_mm
-        )
-
-    def calibrate_extrinsics_from_pose(
-        self,
-        reference_marker_id: int,
-        rvec_camera_marker: "np.ndarray",
-        tvec_camera_marker: "np.ndarray",
-        R_robot_marker: "np.ndarray",
-        t_robot_marker_mm: "np.ndarray",
-    ) -> dict[str, Any]:
-        # Apply extrinsics from a pre-computed marker pose (e.g. averaged
-        # across N frames via bundle PnP) instead of running detect() again.
-        import numpy as np
-
-        marker = ArucoMarker(
-            id=reference_marker_id,
-            corners=np.zeros((4, 2), dtype=np.float32),
-        )
-        marker.rvec_camera = np.asarray(rvec_camera_marker).reshape(3)
-        marker.tvec_camera = np.asarray(tvec_camera_marker).reshape(3)
-        return self._aruco.calibrate_extrinsics(
-            [marker], reference_marker_id, R_robot_marker, t_robot_marker_mm
-        )
-
-    @property
-    def K(self) -> "np.ndarray | None":
-        return self._aruco.K
-
-    @property
-    def dist(self) -> "np.ndarray | None":
-        return self._aruco.dist
-
-    def detect_from_frame(self, frame: "np.ndarray") -> list[ArucoMarker]:
-        return self._aruco.detect_in_frame(frame)
-
-    def is_intrinsics_loaded(self) -> bool:
-        return self._aruco.is_intrinsics_loaded()
-
-    def is_extrinsics_loaded(self) -> bool:
-        return self._aruco.is_extrinsics_loaded()
-
-    def load_calibration(self, path: str | None = None) -> bool:
-        return self._aruco.load(path)
-
-    def save_calibration(self, path: str | None = None) -> None:
-        self._aruco.save(path)
-
-    @property
-    def width(self) -> int:
-        return self._width
-
-    @property
-    def height(self) -> int:
-        return self._height
 
     # Simulation helpers
 
@@ -799,33 +383,38 @@ class UvcCameraVirtual(Camera):
         return dict(self._settings)
 
 
-class UvcCameraVirtualDefinition(DriverDefinition):
+def _add_camera_args(defn: DriverInitArgsDefinition) -> DriverInitArgsDefinition:
+    defn.add_required("device", ArgTypes.String())
+    defn.add_optional("width", ArgTypes.U32(), DEFAULT_WIDTH)
+    defn.add_optional("height", ArgTypes.U32(), DEFAULT_HEIGHT)
+    defn.add_optional("fourcc", ArgTypes.String(), DEFAULT_FOURCC)
+    defn.add_optional("focus", ArgTypes.I32(), -1)  # -1 = no override
+    defn.add_optional("autofocus", ArgTypes.Bool(), False)
+    defn.add_optional("marker_size_mm", ArgTypes.F32(), 32.0)
+    defn.add_optional("dictionary", ArgTypes.String(), DEFAULT_ARUCO_DICTIONARY)
+    defn.add_optional("calibration_path", ArgTypes.String(), "")
+    defn.add_optional("charuco_squares_x", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_X)
+    defn.add_optional("charuco_squares_y", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_Y)
+    defn.add_optional("charuco_square_mm", ArgTypes.F32(), DEFAULT_CHARUCO_SQUARE_MM)
+    defn.add_optional("charuco_marker_mm", ArgTypes.F32(), DEFAULT_CHARUCO_MARKER_MM)
+    return defn
+
+
+class _UvcCameraDefinitionBase(DriverDefinition):
+    _camera_cls: type[_AbstractUvcCamera] = _AbstractUvcCamera
+
     def __init__(self, logger: Logger):
-        super().__init__(UvcCameraVirtual.commands)
+        super().__init__(self._camera_cls.commands)
         self._logger = logger
 
     def get_init_args_definition(self) -> DriverInitArgsDefinition:
-        defn = DriverInitArgsDefinition()
-        defn.add_required("device", ArgTypes.String())
-        defn.add_optional("width", ArgTypes.U32(), DEFAULT_WIDTH)
-        defn.add_optional("height", ArgTypes.U32(), DEFAULT_HEIGHT)
-        defn.add_optional("fourcc", ArgTypes.String(), DEFAULT_FOURCC)
-        defn.add_optional("focus", ArgTypes.I32(), -1)
-        defn.add_optional("autofocus", ArgTypes.Bool(), False)
-        defn.add_optional("marker_size_mm", ArgTypes.F32(), 32.0)
-        defn.add_optional("dictionary", ArgTypes.String(), DEFAULT_ARUCO_DICTIONARY)
-        defn.add_optional("calibration_path", ArgTypes.String(), "")
-        defn.add_optional("charuco_squares_x", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_X)
-        defn.add_optional("charuco_squares_y", ArgTypes.U32(), DEFAULT_CHARUCO_SQUARES_Y)
-        defn.add_optional("charuco_square_mm", ArgTypes.F32(), DEFAULT_CHARUCO_SQUARE_MM)
-        defn.add_optional("charuco_marker_mm", ArgTypes.F32(), DEFAULT_CHARUCO_MARKER_MM)
-        return defn
+        return _add_camera_args(DriverInitArgsDefinition())
 
-    def create(self, args: DriverInitArgs) -> UvcCameraVirtual:
+    def create(self, args: DriverInitArgs) -> _AbstractUvcCamera:
         focus_arg = args.get("focus")
         focus = None if focus_arg == -1 else int(focus_arg)
         cal_path = args.get("calibration_path") or None
-        return UvcCameraVirtual(
+        return self._camera_cls(
             name=args.get_name(),
             logger=self._logger,
             device=args.get("device"),
@@ -843,3 +432,11 @@ class UvcCameraVirtualDefinition(DriverDefinition):
             charuco_square_mm=args.get("charuco_square_mm"),
             charuco_marker_mm=args.get("charuco_marker_mm"),
         )
+
+
+class UvcCameraDefinition(_UvcCameraDefinitionBase):
+    _camera_cls = UvcCamera
+
+
+class UvcCameraVirtualDefinition(_UvcCameraDefinitionBase):
+    _camera_cls = UvcCameraVirtual
