+++ src/evo_lib/perception/__init__.py
@@ -1,3 +1,3 @@
-from evo_lib.perception.aruco import CharucoCalibrationSession
+from evo_lib.perception.aruco import ArucoState, CharucoCalibrationSession
 
-__all__ = ["CharucoCalibrationSession"]
+__all__ = ["ArucoState", "CharucoCalibrationSession"]
